--- src/Files.App/Actions/Display/Group/GroupAscendingAction.cs
@@ -1,44 +1,0 @@
-﻿using CommunityToolkit.Mvvm.ComponentModel;
-using CommunityToolkit.Mvvm.DependencyInjection;
-using Files.App.Contexts;
-using Files.App.Extensions;
-using Files.Shared.Enums;
-using System.ComponentModel;
-using System.Threading.Tasks;
-
-namespace Files.App.Actions
-{
-	internal class GroupAscendingAction : ObservableObject, IToggleAction
-	{
-		private IDisplayPageContext context = Ioc.Default.GetRequiredService<IDisplayPageContext>();
-
-		public string Label { get; } = "Ascending".GetLocalizedResource();
-
-		public bool IsOn => context.GroupDirection is SortDirection.Ascending;
-		public bool IsExecutable => context.GroupOption is not GroupOption.None;
-
-		public GroupAscendingAction()
-		{
-			context.PropertyChanged += Context_PropertyChanged;
-		}
-
-		public Task ExecuteAsync()
-		{
-			context.GroupDirection = SortDirection.Ascending;
-			return Task.CompletedTask;
-		}
-
-		private void Context_PropertyChanged(object? sender, PropertyChangedEventArgs e)
-		{
-			switch (e.PropertyName)
-			{
-				case nameof(IDisplayPageContext.GroupOption):
-					OnPropertyChanged(nameof(IsExecutable));
-					break;
-				case nameof(IDisplayPageContext.GroupDirection):
-					OnPropertyChanged(nameof(IsOn));
-					break;
-			}
-		}
-	}
-}
