--- src/Files.App/Actions/Display/Group/GroupByDateCreatedAction.cs
@@ -1,12 +1,0 @@
-﻿using Files.App.Extensions;
-using Files.Shared.Enums;
-
-namespace Files.App.Actions
-{
-	internal class GroupByDateCreatedAction : GroupByAction
-	{
-		protected override GroupOption GroupOption { get; } = GroupOption.DateCreated;
-
-		public override string Label { get; } = "DateCreated".GetLocalizedResource();
-	}
-}
