--- src/Files.App/Actions/Display/Group/GroupBySizeAction.cs
@@ -1,12 +1,0 @@
-﻿using Files.App.Extensions;
-using Files.Shared.Enums;
-
-namespace Files.App.Actions
-{
-	internal class GroupBySizeAction : GroupByAction
-	{
-		protected override GroupOption GroupOption { get; } = GroupOption.Size;
-
-		public override string Label { get; } = "Size".GetLocalizedResource();
-	}
-}
