--- src/Files.App/Actions/Display/Group/GroupByTagAction.cs
@@ -1,12 +1,0 @@
-﻿using Files.App.Extensions;
-using Files.Shared.Enums;
-
-namespace Files.App.Actions
-{
-	internal class GroupByTagAction : GroupByAction
-	{
-		protected override GroupOption GroupOption { get; } = GroupOption.FileTag;
-
-		public override string Label { get; } = "FileTags".GetLocalizedResource();
-	}
-}
