--- src/Files.App/Actions/Display/Group/GroupByTypeAction.cs
@@ -1,12 +1,0 @@
-﻿using Files.App.Extensions;
-using Files.Shared.Enums;
-
-namespace Files.App.Actions
-{
-	internal class GroupByTypeAction : GroupByAction
-	{
-		protected override GroupOption GroupOption { get; } = GroupOption.FileType;
-
-		public override string Label { get; } = "Type".GetLocalizedResource();
-	}
-}
