--- src/Files.App/Actions/Display/Group/GroupDescendingAction.cs
@@ -1,44 +1,0 @@
-﻿using CommunityToolkit.Mvvm.ComponentModel;
-using CommunityToolkit.Mvvm.DependencyInjection;
-using Files.App.Contexts;
-using Files.App.Extensions;
-using Files.Shared.Enums;
-using System.ComponentModel;
-using System.Threading.Tasks;
-
-namespace Files.App.Actions
-{
-	internal class GroupDescendingAction : ObservableObject, IToggleAction
-	{
-		private IDisplayPageContext context = Ioc.Default.GetRequiredService<IDisplayPageContext>();
-
-		public string Label { get; } = "Descending".GetLocalizedResource();
-
-		public bool IsOn => context.GroupDirection is SortDirection.Descending;
-		public bool IsExecutable => context.GroupOption is not GroupOption.None;
-
-		public GroupDescendingAction()
-		{
-			context.PropertyChanged += Context_PropertyChanged;
-		}
-
-		public Task ExecuteAsync()
-		{
-			context.GroupDirection = SortDirection.Descending;
-			return Task.CompletedTask;
-		}
-
-		private void Context_PropertyChanged(object? sender, PropertyChangedEventArgs e)
-		{
-			switch (e.PropertyName)
-			{
-				case nameof(IDisplayPageContext.GroupOption):
-					OnPropertyChanged(nameof(IsExecutable));
-					break;
-				case nameof(IDisplayPageContext.GroupDirection):
-					OnPropertyChanged(nameof(IsOn));
-					break;
-			}
-		}
-	}
-}
