--- src/Files.App/Actions/Display/Sort/SortAscendingAction.cs
@@ -1,36 +1,0 @@
-﻿using CommunityToolkit.Mvvm.ComponentModel;
-using CommunityToolkit.Mvvm.DependencyInjection;
-using Files.App.Contexts;
-using Files.App.Extensions;
-using Files.Shared.Enums;
-using System.ComponentModel;
-using System.Threading.Tasks;
-
-namespace Files.App.Actions
-{
-	internal class SortAscendingAction : ObservableObject, IToggleAction
-	{
-		private IDisplayPageContext context = Ioc.Default.GetRequiredService<IDisplayPageContext>();
-
-		public string Label { get; } = "Ascending".GetLocalizedResource();
-
-		public bool IsOn => context.SortDirection is SortDirection.Ascending;
-
-		public SortAscendingAction()
-		{
-			context.PropertyChanged += Context_PropertyChanged;
-		}
-
-		public Task ExecuteAsync()
-		{
-			context.SortDirection = SortDirection.Ascending;
-			return Task.CompletedTask;
-		}
-
-		private void Context_PropertyChanged(object? sender, PropertyChangedEventArgs e)
-		{
-			if (e.PropertyName is nameof(IDisplayPageContext.SortDirection))
-				OnPropertyChanged(nameof(IsOn));
-		}
-	}
-}
