--- src/Files.App/Actions/Display/Sort/SortByDateCreatedAction.cs
@@ -1,12 +1,0 @@
-﻿using Files.App.Extensions;
-using Files.Shared.Enums;
-
-namespace Files.App.Actions
-{
-	internal class SortByDateCreatedAction : SortByAction
-	{
-		protected override SortOption SortOption { get; } = SortOption.DateCreated;
-
-		public override string Label { get; } = "DateCreated".GetLocalizedResource();
-	}
-}
