--- src/Files.App/Actions/Display/Sort/SortBySizeAction.cs
@@ -1,12 +1,0 @@
-﻿using Files.App.Extensions;
-using Files.Shared.Enums;
-
-namespace Files.App.Actions
-{
-	internal class SortBySizeAction : SortByAction
-	{
-		protected override SortOption SortOption { get; } = SortOption.Size;
-
-		public override string Label { get; } = "Size".GetLocalizedResource();
-	}
-}
