--- src/Files.App/Actions/Display/Sort/SortBySyncStatusAction.cs
@@ -1,15 +1,0 @@
-﻿using Files.App.Contexts;
-using Files.App.Extensions;
-using Files.Shared.Enums;
-
-namespace Files.App.Actions
-{
-	internal class SortBySyncStatusAction : SortByAction
-	{
-		protected override SortOption SortOption { get; } = SortOption.SyncStatus;
-
-		public override string Label { get; } = "SyncStatus".GetLocalizedResource();
-
-		protected override bool GetIsExecutable(ContentPageTypes pageType) => pageType is ContentPageTypes.CloudDrive;
-	}
-}
