--- src/Files.App/Actions/Display/Sort/SortByTypeAction.cs
@@ -1,12 +1,0 @@
-﻿using Files.App.Extensions;
-using Files.Shared.Enums;
-
-namespace Files.App.Actions
-{
-	internal class SortByTypeAction : SortByAction
-	{
-		protected override SortOption SortOption { get; } = SortOption.FileType;
-
-		public override string Label { get; } = "Type".GetLocalizedResource();
-	}
-}
