+++ src/Files.App/Actions/Display/ToggleSortDirectoriesAlongsideFilesAction.cs
